+++ src/utils/sanitizeCallbackUrl.ts
@@ -9,5 +9,8 @@ export const sanitizeCallbackUrl = (
 
   const questionMark = Array.from(params.values()).length > 0 ? '?' : '';
 
-  return `${url.origin}${url.pathname}${questionMark}${params.toString()}`;
+  const pathname =
+    url.pathname === '/' && !targetURL.endsWith('/') ? '' : url.pathname;
+
+  return `${url.origin}${pathname}${questionMark}${params.toString()}`;
 };